refactor(socialaccount/providers): Split up `base.py`
