+++ allauth/socialaccount/providers/base/provider.py
@@ -1,25 +1,6 @@
 from allauth.account.models import EmailAddress
 from allauth.socialaccount import app_settings
-
-from ..adapter import get_adapter
-
-
-class AuthProcess(object):
-    LOGIN = "login"
-    CONNECT = "connect"
-    REDIRECT = "redirect"
-
-
-class AuthAction(object):
-    AUTHENTICATE = "authenticate"
-    REAUTHENTICATE = "reauthenticate"
-    REREQUEST = "rerequest"
-
-
-class AuthError(object):
-    UNKNOWN = "unknown"
-    CANCELLED = "cancelled"  # Cancelled on request of user
-    DENIED = "denied"  # Denied by server
+from allauth.socialaccount.adapter import get_adapter
 
 
 class ProviderException(Exception):
